Remove inner functions and types
