+++ src/main/kotlin/com/github/derg/transpiler/phases/parser/ParserStatements.kt
@@ -21,13 +21,8 @@ private fun merge(name: Name, operator: SymbolType, rhs: Expression): Assignment
 /**
  * Parses a single statement from the token stream.
  */
-fun statementParserOf(): Parser<Statement> =
-    ParserPattern(::statementPatternOf) { it }
-
-private fun statementPatternOf(): Parser<Statement> = ParserAnyOf(
+fun statementParserOf(): Parser<Statement> = ParserAnyOf(
     variableParserOf(),
-    functionParserOf(),
-    typeParserOf(),
     assignmentParserOf(),
     branchParserOf(),
     invokeParserOf(),
@@ -118,77 +113,3 @@ private fun returnOutcomeOf(values: Parsers): Control.Return
 // TODO: Not a correct implementation of the parser - must also function with error handling
 private fun invokeParserOf(): Parser<Statement> =
     ParserPattern(::functionCallParserOf) { Control.Invoke(it) }
-
-/**
- * Parses a variable definition from the token stream.
- */
-fun variableParserOf(): Parser<Statement> =
-    ParserPattern(::variablePatternOf, ::variableOutcomeOf)
-
-private fun variablePatternOf() = ParserSequence(
-    "visibility" to visibilityParserOf(),
-    "assignability" to assignabilityParserOf(),
-    "mutability" to mutabilityParserOf(),
-    "name" to ParserName(),
-    "op" to ParserSymbol(SymbolType.ASSIGN),
-    "value" to expressionParserOf(),
-)
-
-private fun variableOutcomeOf(values: Parsers) = Definition.Variable(
-    name = values["name"],
-    type = null,
-    value = values["value"],
-    visibility = values["visibility"],
-    mutability = values["mutability"],
-    assignability = values["assignability"],
-)
-
-/**
- * Parses a function definition from the token stream.
- */
-fun functionParserOf(): Parser<Statement> =
-    ParserPattern(::functionPatternOf, ::functionOutcomeOf)
-
-private fun functionPatternOf() = ParserSequence(
-    "visibility" to visibilityParserOf(),
-    "fun" to ParserSymbol(SymbolType.FUN),
-    "name" to ParserName(),
-    "open_parenthesis" to ParserSymbol(SymbolType.OPEN_PARENTHESIS),
-    "parameters" to ParserRepeating(parameterParserOf(), ParserSymbol(SymbolType.COMMA)),
-    "close_parenthesis" to ParserSymbol(SymbolType.CLOSE_PARENTHESIS),
-    "error" to ParserOptional(nameParserOf(SymbolType.COLON)),
-    "value" to ParserOptional(nameParserOf(SymbolType.ARROW)),
-    "open_brace" to ParserSymbol(SymbolType.OPEN_BRACE),
-    "statements" to ParserRepeating(statementParserOf()),
-    "close_brace" to ParserSymbol(SymbolType.CLOSE_BRACE),
-)
-
-private fun functionOutcomeOf(values: Parsers) = Definition.Function(
-    name = values["name"],
-    valueType = values["value"],
-    errorType = values["error"],
-    parameters = values["parameters"],
-    visibility = values["visibility"],
-    statements = values["statements"],
-)
-
-/**
- * Parses a type definition from the token stream.
- */
-fun typeParserOf(): Parser<Statement> =
-    ParserPattern(::typePatternOf, ::typeOutcomeOf)
-
-private fun typePatternOf() = ParserSequence(
-    "visibility" to visibilityParserOf(),
-    "type" to ParserSymbol(SymbolType.TYPE),
-    "name" to ParserName(),
-    "open_brace" to ParserSymbol(SymbolType.OPEN_BRACE),
-    "properties" to ParserRepeating(propertyParserOf()),
-    "close_brace" to ParserSymbol(SymbolType.CLOSE_BRACE),
-)
-
-private fun typeOutcomeOf(values: Parsers) = Definition.Type(
-    name = values["name"],
-    visibility = values["visibility"],
-    properties = values["properties"],
-)
